Create Theseus Master CAD Redesign.zip
Main CAD based on reference planes and a single point (origin).

All assemblies reference equations for part heights and distances between parts and reference planes. Any change to part heights should be done through equations, as it'll automatically update the main assembly.

Important to note that there is a set of dimensional equations in the main assembly that dictate the distance for the fluids system in each assembly section. To view these values, turn on view mode (eye ball in the center dash) and select show dimensions under annotations,